requirements.txt
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-bootstrap4==2.0.1
django-phonenumber-field==4.0.0
gunicorn==20.0.4
idna==2.9
phonenumbers==8.12.5
pytz==2020.1
requests==2.23.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
